modal-backend/pyproject.toml
[project]
name = "modal-backend"
version = "0.1.0"
description = "OpenTrace video rendering backend on Modal"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.3.0.post1",
    "pillow>=12.0.0",
]

[tool.uv]
dev-dependencies = []
